requirements.txt
aioconsole==0.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncio-atexit==1.0.1
attrs==25.1.0
autogen-agentchat==0.4.0.dev13
autogen-core==0.4.0.dev13
autogen-ext==0.4.0.dev13
azure-cognitiveservices-speech==1.42.0
azure-core==1.32.0
azure-identity==1.19.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Deprecated==1.2.18
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.4.0
jiter==0.8.2
jsonref==1.1.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
openai==1.61.1
opentelemetry-api==1.27.0
pillow==11.1.0
portalocker==2.10.1
propcache==0.2.1
protobuf==4.25.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
